// client/src/api/generated-client.ts
export interface EventDto {
    id?: string | undefined;
    title?: string;
    description?: string;
    date?: Date;
    time?: string;
    location?: string;
    category?: string;
    status?: string;
}